requirements.txt
dominate==2.6.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask==2.0.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.2
pandas==1.4.0
pip==21.1.2
plotly==5.5.0
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
setuptools==57.0.0
six==1.16.0
SQLAlchemy==1.4.29
tenacity==8.0.1
urllib3==1.26.8
validate-email==1.3
visitor==0.1.3
Werkzeug==2.0.2
wheel==0.36.2
